pyproject.toml
[tool.poetry]
name = "nonebot-plugin-essence-message"
version = "0.1.0"
description = "Used to organize essence messages"
license = "MIT"
authors = ["beis <havaagoodtime@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.32.3"
httpx = "^0.27.0"
nonebot2 = "^2.3.2"
nonebot-adapter-onebot = "^2.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"